Time: 0 ms (100%), Space: 8.6 MB (53.81%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -10,6 +10,6 @@ public:
             if (rev > INT_MAX || rev < INT_MIN) 
                 return 0;
         }
-        return (int)rev;
+        return rev;
     }
 };
